base/requirements.txt
Django==1.10.6
djangorestframework==3.5.4
Pillow==4.0.0
pytest==3.0.6
pytest-cov==2.4.0
pytest-django==3.1.2
